Code refactoring + tests
