+++ vuejs_project/src/App.vue
@@ -18,7 +18,7 @@
     <RegisterModal ref="myRegisterModal" />
     <LoginModal ref="myLoginModal" />
     <GroupManagementModal ref="myGroupManagementModal" />
-    <router-view style="padding-bottom:50px" :theme="theme"/>
+    <router-view style="padding-bottom:50px" :theme="theme" ref="portal" />
   </div>
 </template>
 
@@ -27,7 +27,10 @@ import RegisterModal from './components/auth/RegisterModal.vue'
 import LoginModal from './components/auth/LoginModal.vue'
 import GroupManagementModal from './components/auth/GroupManagementModal.vue'
 import { Component, Vue } from 'vue-property-decorator'
-import { THEME } from './config'
+import { THEME, COMMUNICATION_TYPE } from './config'
+import JSEncrypt from 'jsencrypt'
+import DesignBoard from './views/DesignBoard.vue'
+import ConceptionGrid from './components/conception/ConceptionGrid.vue'
 
 @Component({
   components: {
@@ -38,10 +41,143 @@ import { THEME } from './config'
 })
 export default class App extends Vue {
   // Dark or light
-  private theme = THEME;
+  private theme = THEME
+  private connection: WebSocket | null = null
+  public shouldReload = 2 // Backend send designerdata twice in short time
+  private backendUrl: string | undefined = process.env.VUE_APP_BACKEND_URL
+  private encryptForBackend = new JSEncrypt()
+  private decryptForFrontend = new JSEncrypt()
+  private dataReceiving = ''
+
+  mounted () {
+    this.$nextTick(function () {
+      // If the node environment is test, we populate the toolbar of fake components for tests
+      if (process.env.NODE_ENV === 'test' || !this.backendUrl) {
+        if (this.$refs.portal instanceof DesignBoard) {
+          (this.$refs.portal as DesignBoard).sendBlankDesignerData()
+        }
+      } else {
+        this.encryptForBackend.setPublicKey(process.env.VUE_APP_BACKEND_PUBLIC_KEY)
+        this.connect(3, this.backendUrl)
+      }
+    })
+  }
+
+  /**
+   * Connect to backend by WebSocket
+   * And initialize listeners for communication
+   * @param connectionTries the maximum number of connections we want tries before give up
+   * @param url the backend url. Ex: ws://localhost:5001
+   * @public
+   */
+  connect (connectionTries: number, url: string): void {
+    const treatMessage = (msg: string) => {
+      let res = ''
+      msg.split(',').forEach(el => { res += this.decryptForFrontend.decrypt(el) })
+      const data = JSON.parse(res)
+      switch (data.type) {
+        case 'debug':
+          if (this.$refs.portal instanceof DesignBoard) {
+            (this.$refs.portal as DesignBoard).sendMessage(data)
+          }
+          break
+        case 'designer':
+          if (this.shouldReload > 0 || COMMUNICATION_TYPE === 'DIRECT') {
+            if (this.$refs.portal instanceof DesignBoard) {
+              (this.$refs.portal as DesignBoard).sendDesignerData(data)
+            }
+            --this.shouldReload
+            // if too many time is spend (1.5s) before the next designer data, close the opportunnity to reload
+            setTimeout(() => { this.shouldReload = 0 }, 1500)
+          }
+          break
+        case 'error':
+        case 'errors':
+          console.error(data.type, data)
+          break
+        case 'online':
+          console.log('Count of client connected: ' + data.count)
+          break
+        case 'status':
+          console.log('Message type "' + data.type + '".')
+          console.log(data)
+          break
+        case 'traffic':
+          if (this.$refs.portal instanceof DesignBoard) {
+            ((this.$refs.portal as DesignBoard).$children[1] as ConceptionGrid).setTraffic(data.body)
+          }
+          break
+        default:
+          console.warn('Message type "' + data.type + '" not treated.')
+          break
+      }
+    }
+    const giveFrontendPublicKey = () => {
+      const msg = { type: 'key', body: this.decryptForFrontend.getPublicKey() }
+      this.sendMessageToBackend([JSON.stringify(msg)])
+    }
+    console.log('Starting connection to WebSocket Server...')
+    this.connection = new WebSocket(url)
+    this.connection.addEventListener('error', e => {
+      // readyState === 3 is CLOSED
+      if ((e.target as WebSocket).readyState === 3) {
+        if (connectionTries > 0) {
+          setTimeout(() => this.connect(connectionTries - 1, url), 1000)
+        } else {
+          console.error('Maximum number of connection trials has been reached')
+        }
+      } else {
+        console.error('Websocket error: ' + event?.target)
+      }
+    })
+    this.connection.onopen = function () {
+      console.log('Successfully connected to the echo websocket server...')
+      giveFrontendPublicKey()
+    }
+    this.connection.onmessage = function (event) {
+      treatMessage(decodeURIComponent(event.data))
+    }
+  }
+
+  /**
+   * Call by a listener in addComponent.ts
+   * Will send a message to backend for trigger clicked component
+   * @param data
+   * @public
+   */
+  sendMessageToBackend (data: Array<string>): void {
+    if (this.connection !== null) {
+      data.forEach(el => {
+        if (this.connection !== null) {
+          this.shouldReload = 2
+          if (el.length <= 125) {
+            this.connection.send(this.encryptForBackend.encrypt(el))
+          } else {
+            let offset = 0
+            let res = ''
+            while (offset < el.length) {
+              const size = Math.min(125, el.length - offset)
+              res += this.encryptForBackend.encrypt(el.substring(offset, offset + size))
+              offset += size
+              if (offset < el.length) {
+                res += ','
+              }
+            }
+            this.connection.send(res)
+          }
+        }
+      })
+    } else {
+      console.log('Data received but no connection to send the message found.')
+    }
+  }
 
   get currentRoute () { return this.$route.path }
 
+  /**
+   * Open AUthentification modals
+   * @public
+   */
   openModal (modal: 'register' | 'login' | 'group'): void {
     switch (modal) {
       case 'register':
